pyproject.toml
[project]
name = "exploration-202604"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-openai>=1.1.12",
    "load-dotenv>=0.1.0",
    "numpy>=2.4.4",
]
